Moved study logic out of main.rs
