+++ src/scenes/mod.rs
@@ -1,3 +1,4 @@
 // src/scenes/mod.rs
 
 pub mod main_menu;
+pub mod studying;
